requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
geoalchemy2==0.14.2
psycopg2-binary==2.9.9
google-cloud-storage==2.10.0
google-auth==2.25.2
python-dotenv==1.0.0
requests==2.31.0
mercadopago==2.2.0
httpx[http2]>=0.26,<0.29
websockets>=11,<16
boto3==1.29.7
fpdf2>=2.7.8
qrcode>=7.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Celery + Redis
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Data processing
scipy==1.12.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
h3==4.1.2

# Logging
python-json-logger==2.0.7

# Utilities
pydantic>=2.11.7,<3
pydantic-settings>=2.2,<3
email-validator>=2.1.0
Pillow==10.2.0

# Google Earth Engine (opcional, para VAE + recovery tasks)
earthengine-api>=1.6.12
